+++ tools/ice-devtools/utils/build-combined-demo.js
@@ -1,3 +1,5 @@
+/* Deprecated */
+/* eslint-disable */
 const fs = require('fs');
 const path = require('path');
 const rimraf = require('rimraf');
@@ -92,7 +94,7 @@ function generateEntryJS(demos) {
   const tempPath = getTempPath();
   const jsPath = path.join(tempPath, `component-index.js`);
 
-  const jsTemplateContent = compileTemplateContent({ 
+  const jsTemplateContent = compileTemplateContent({
     demos: demos.map((demo) => {
       return {
         path: formatPathForWin(demo.filePath)