+++ tools/ice-devtools/utils/get-demos.js
@@ -1,3 +1,5 @@
+/* Deprecated */
+/* eslint-disable */
 const { readdirSync, readFileSync, existsSync } = require('fs');
 const { join } = require('path');
 
@@ -15,18 +17,20 @@ module.exports = function getDemos(projectDir) {
       const filePath = join(demoPath, filename);
       const content = readFileSync(filePath, 'utf-8');
 
-      const { meta, highlightedCode, content: markdownContent } = parseMarkdownParts(content);
+      const { meta, highlightedCode, content: markdownContent } = parseMarkdownParts(content, {
+        sliceCode: true
+      });
 
       filename = filename.replace(/\.md$/, '');
       const href = `/preview/?demo=${filename}`;
 
-      return { 
-        href, 
-        filename, 
-        filePath, 
-        ...meta, 
-        highlightedCode, 
-        markdownContent, 
+      return {
+        href,
+        filename,
+        filePath,
+        ...meta,
+        highlightedCode,
+        markdownContent,
       };
     }).sort(function (a, b) {
       return a.order - b.order;