+++ tools/ice-devtools/utils/transform.js
@@ -2,10 +2,12 @@ module.exports = (files, data, done) => {
   Object.keys(files).forEach((filename) => {
     let newFilename = filename;
 
+    /* eslint-disable-next-line no-useless-escape */
     if (/[_\.]/.test(filename)) {
+      /* eslint-disable-next-line no-useless-escape */
       newFilename = filename.replace(/[_\.]_className__/, data.className);
     }
-    if (/_gitignore/.test(filename)) {
+    if (/_gitignore/.test(filename) && !data.skipGitIgnore) {
       newFilename = filename.replace(/_gitignore/, '.gitignore');
     }
     if (/_package.json/.test(filename)) {