+++ tools/ice-scripts/bin/ice.js
@@ -1,17 +1,18 @@
 #!/usr/bin/env node
 
-'use strict';
 
 const program = require('commander');
 
 const packageInfo = require('../package.json');
 
 console.log(packageInfo.name, packageInfo.version);
+
 program
   .version(packageInfo.version)
   .usage('<command> [options]')
   .command('build', 'build project')
-  .command('dev', 'start server');
+  .command('dev', 'start server')
+  .command('init', 'init project by scaffold');
 
 program.parse(process.argv);