+++ tools/ice-scripts/lib/config/getResolveAlias.js
@@ -1,12 +1,15 @@
 const path = require('path');
-const colors = require('chalk');
+
+const paths = require('./paths');
+const pkgData = require('./packageJson');
+const log = require('../utils/log');
 
 module.exports = function getResolveAlias(buildConfig) {
   let alias = {};
 
   if (buildConfig.uniteBaseComponent) {
     // 统一基础组件包：@ali/ice, @alife/next, @icedesign/base -> @icedesign/base
-    console.log(colors.green('Info:'), 'uniteBaseComponent 开启，基础包统一到 @icedesign/base');
+    log.info('uniteBaseComponent 开启，基础包统一到 @icedesign/base');
     alias = {
       '@ali/ice/global.scss': '@icedesign/base/reset.scss',
       '@ali/ice/lib/row$': '@icedeisign/base/lib/_components/@alife/next-grid/lib/row.js',
@@ -24,8 +27,12 @@ module.exports = function getResolveAlias(buildConfig) {
       '@alife/next/variables.scss': '@icedesign/base/variables.scss',
       '@alife/next/lib/core/index.scss': '@icedesign/base/lib/core/index.scss',
 
-      '@alife/next': '@icedesign/base'
+      '@alife/next': '@icedesign/base',
     };
+  }
+
+  if (pkgData.type === 'component') {
+    alias[pkgData.name] = path.resolve(paths.appDirectory, 'src/index');
   }
 
   return alias;