+++ tools/ice-scripts/lib/plugins/check-ice-components-dep.js
@@ -1,29 +1,26 @@
+/* eslint-disable import/no-dynamic-require */
 /**
  * 检测 ICE 组件的依赖问题
  *  - 部分代码直接借鉴了 duplicate-package-checker-webpack-plugin
  */
 const path = require('path');
 const findRoot = require('find-root');
-const colors = require('chalk');
 const semver = require('semver');
 const bizComponentsVersion = require('./bizComponentsVersion');
 const deprecatedComponents = require('./deprecatedComponents');
 
+const log = require('../utils/log');
+
 const depModules = {};
 
 module.exports = class CheckDepsPlugin {
-
   constructor(options) {
     this.pkg = options.pkg || {};
   }
 
   apply(compiler) {
     compiler.hooks.emit.tapAsync('CheckDepsPlugin', (compilation, callback) => {
-      const projectPath = compilation.compiler.context;
-      const npmInfos = [];
-
       compilation.modules.forEach((module) => {
-
         if (!module.resource) {
           return;
         }
@@ -36,7 +33,6 @@ module.exports = class CheckDepsPlugin {
         }
 
         const pkg = closestPackage.package;
-        const packagePath = closestPackage.path;
 
         if (!depModules[pkg.name]) {
           depModules[pkg.name] = [pkg.version];
@@ -51,14 +47,14 @@ module.exports = class CheckDepsPlugin {
       Object.keys(depModules).forEach((moduleName) => {
         const versions = depModules[moduleName];
         if (versions.length > 1) {
-          console.log(colors.yellow('Warning: '), `项目依赖了 ${moduleName} 的两个版本，`, colors.green(versions));
+          log.warn(`项目依赖了 ${moduleName} 的两个版本，`, versions);
         }
       });
 
       // 2. 多份基础组件
-      const baseComponentDeps = ['@icedesign/base', '@alife/next', '@ali/ice'].filter(name => depModules[name]);
+      const baseComponentDeps = ['@icedesign/base', '@alife/next', '@ali/ice'].filter((name) => depModules[name]);
       if (baseComponentDeps.length > 1) {
-        console.log(colors.red('Error: '), `项目依赖了多份基础组件 ${baseComponentDeps}，建议通过配置 buildConfig.uniteBaseComponent 优化`);
+        log.warn(`项目依赖了多份基础组件 ${baseComponentDeps}，建议通过配置 buildConfig.uniteBaseComponent 优化`);
       }
 
       // 3. 业务组件与基础组件的版本对应关系
@@ -69,14 +65,14 @@ module.exports = class CheckDepsPlugin {
       if (depFeNext && !depFdNext) {
         // 只依赖了 0.x 的项目应该使用 0.x 的业务组件
         Object.keys(depModules).forEach((moduleName) => {
-          checkBizComponentVersion(moduleName, depModules[moduleName][0], '0.x')
+          checkBizComponentVersion(moduleName, depModules[moduleName][0], '0.x');
         });
       }
 
       if (depFdNext && !depFeNext) {
         // 只依赖了 1.x 的项目应该使用 1.x 的业务组件
         Object.keys(depModules).forEach((moduleName) => {
-          checkBizComponentVersion(moduleName, depModules[moduleName][0], '1.x')
+          checkBizComponentVersion(moduleName, depModules[moduleName][0], '1.x');
         });
       }
 
@@ -84,15 +80,14 @@ module.exports = class CheckDepsPlugin {
       Object.keys(depModules).forEach((moduleName) => {
         const deprecatedMsg = deprecatedComponents[moduleName];
         if (deprecatedMsg) {
-          console.log(colors.yellow('Warning: '),deprecatedMsg);
+          log.warn(deprecatedMsg);
         }
       });
 
       callback();
     });
   }
-
-}
+};
 
 // Get closest package definition from path
 function getClosestPackage(modulePath) {
@@ -102,7 +97,7 @@ function getClosestPackage(modulePath) {
   // Catch findRoot or require errors
   try {
     root = findRoot(modulePath);
-    pkg = require(path.join(root, "package.json"));
+    pkg = require(path.join(root, 'package.json'));
   } catch (e) {
     return null;
   }
@@ -112,12 +107,12 @@ function getClosestPackage(modulePath) {
   // https://github.com/jsdnxx/find-root/issues/2
   // https://github.com/date-fns/date-fns/issues/264#issuecomment-265128399
   if (!pkg.name) {
-    return getClosestPackage(path.resolve(root, ".."));
+    return getClosestPackage(path.resolve(root, '..'));
   }
 
   return {
     package: pkg,
-    path: root
+    path: root,
   };
 }
 
@@ -127,17 +122,16 @@ function checkBizComponentVersion(npmName, npmVersion, baseVersion) {
     return;
   }
 
-  const version = npmVersion;
   const semverVersion = bizComponentsVersion[npmName][baseVersion];
 
   if (!semverVersion) {
     // 没有对应的（未升级）
-    console.log(colors.yellow('Warning: '), `${npmName} 暂时没有符合基础组件 ${baseVersion} 的版本，建议联系 ICE 团队协助升级`);
+    log.warn(`${npmName} 暂时没有符合基础组件 ${baseVersion} 的版本，建议联系 ICE 团队协助升级`);
   }
 
   if (!semver.satisfies(npmVersion, semverVersion)) {
     // 不符合版本
-    console.log(colors.red('Error: '), `项目使用的基础组件版本是 ${baseVersion}，业务组件 ${npmName}@${npmVersion} 不符合版本要求 ${semverVersion}，建议选择正确的组件版本`);
+    log.warn(`项目使用的基础组件版本是 ${baseVersion}，业务组件 ${npmName}@${npmVersion} 不符合版本要求 ${semverVersion}，建议选择正确的组件版本`);
   }
 }