+++ tools/iceworks/app/main/template/utils/extractTarball.js
@@ -8,6 +8,7 @@ const tar = require('tar');
 const logger = require('../../logger');
 const { DetailError } = require('../../error-handler');
 const autoRetry = require('../../utils/autoRetry');
+const alilog = require('../../alilog');
 
 /**
  * 将 tarbar 下的内容下载到指定目录，同时做路径转换
@@ -44,6 +45,17 @@ function extractTarball(
         progressFunc(state);
       })
       .on('error', (error = {}) => {
+        alilog.report(
+          {
+            type: 'download-tarball-error',
+            msg: error.message,
+            stack: error.stack,
+            data: {
+              url: tarballURL,
+            },
+          },
+          'error'
+        );
         reject(new DetailError(`链接 ${tarballURL} 请求失败`, {
           message: error.message,
           stack: error.stack,
@@ -117,10 +129,8 @@ function extractTarball(
 
 // 超时自动重试
 const retryCount = 2;
-const retryExtractTarball = autoRetry(extractTarball, retryCount, (err) => {
-  if (err.metadata && err.metadata.message !== 'ETIMEDOUT') {
-    throw (err);
-  }
-});
-
-module.exports = retryExtractTarball;
+module.exports = autoRetry(
+  extractTarball,
+  retryCount,
+  (err) => err.metadata && err.metadata.message === 'ETIMEDOUT'
+);